+++ cli/kata-check_arm64_test.go
@@ -198,3 +198,7 @@ foo     : bar
 		}
 	}
 }
+
+func TestSetCPUtype(t *testing.T) {
+	testSetCPUTypeGeneric(t)
+}
