+++ cli/kata-env.go
@@ -294,6 +294,8 @@ func getHypervisorInfo(config oci.RuntimeConfig) HypervisorInfo {
 }
 
 func getEnvInfo(configFile string, config oci.RuntimeConfig) (env EnvInfo, err error) {
+	setCPUtype()
+
 	meta := getMetaInfo()
 
 	runtime := getRuntimeInfo(configFile, config)
